Wherever you have cache, you also have cache clearing problems. :(
