+++ src-electron/db/query-struct.js
@@ -1,0 +1,44 @@
+/**
+ *
+ *    Copyright (c) 2021 Silicon Labs
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+/**
+ * This module provides queries for enums.
+ *
+ * @module DB API: zcl database access
+ */
+const dbApi = require('./db-api')
+const dbMapping = require('./db-mapping')
+const dbCache = require('./db-cache')
+
+async function selectStructById(db, id) {
+  return dbApi
+    .dbGet(db, 'SELECT STRUCT_ID, NAME FROM STRUCT WHERE STRUCT_ID = ?', [id])
+    .then(dbMapping.map.struct)
+}
+
+async function selectStructByName(db, name, packageId) {
+  return dbApi
+    .dbGet(
+      db,
+      'SELECT STRUCT_ID, NAME FROM STRUCT WHERE NAME = ? AND PACKAGE_REF = ? ORDER BY NAME',
+      [name, packageId]
+    )
+    .then(dbMapping.map.struct)
+}
+
+exports.selectStructById = selectStructById
+exports.selectStructByName = selectStructByName
